+++ src/main/java/hometoogether/hometoogether/domain/pose/domain/PoseType.java
@@ -1,4 +1,4 @@
-package hometoogether.hometoogether.domain.pose;
+package hometoogether.hometoogether.domain.pose.domain;
 
 public enum PoseType {
     PHOTO, VIDEO
